fix: 인기 검색어 조회 기능 오류 개선
