+++ src/main/java/com/example/CacheBoost/domain/searchkeyword/repository/SearchKeywordRepository.java
@@ -1,6 +1,7 @@
 package com.example.CacheBoost.domain.searchkeyword.repository;
 
 import com.example.CacheBoost.domain.searchkeyword.entity.SearchKeyword;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
